+++ examplepatches/humzakh/HK-Morphe-Patches/extensions/syncforreddit/stubs/src/main/java/com/laurencedawson/reddit_sync/singleton/SettingsSingleton.java
@@ -1,0 +1,11 @@
+package com.laurencedawson.reddit_sync.singleton;
+
+public class SettingsSingleton {
+    public static class Settings {
+        public boolean postsLongPress;
+    }
+
+    public static Settings x() {
+        return null;
+    }
+}