+++ examplepatches/humzakh/HK-Morphe-Patches/patches/src/main/kotlin/app/template/patches/reddit/customclients/sync/syncforreddit/fix/redditvideolink/Fingerprints.kt
@@ -1,0 +1,9 @@
+package app.template.patches.reddit.customclients.sync.syncforreddit.fix.redditvideolink
+
+import app.morphe.patcher.Fingerprint
+
+internal val linkHandlerOpenLinkFingerprint = Fingerprint(
+    strings = listOf("Opening strange reddit video link", "Opening link: ")
+)
+
+